+++ Code/GraphMol/FileParsers/test1.cpp
@@ -954,7 +954,6 @@ void testDblBondStereochem() {
     m1 = MolFileToMol(fName);
     TEST_ASSERT(m1);
     TEST_ASSERT(m1->getBondWithIdx(0)->getStereo() == Bond::STEREOANY);
-    TEST_ASSERT(m1->getBondWithIdx(0)->getBondDir() == Bond::EITHERDOUBLE);
     delete m1;
   }
 
@@ -3180,8 +3179,7 @@ void testIssue3375684() {
     RWMol *m = MolFileToMol(fName);
 
     TEST_ASSERT(m->getBondBetweenAtoms(6, 7)->getBondType() == Bond::DOUBLE);
-    TEST_ASSERT(m->getBondBetweenAtoms(6, 7)->getBondDir() ==
-                Bond::EITHERDOUBLE);
+    TEST_ASSERT(m->getBondBetweenAtoms(6, 7)->getStereo() == Bond::STEREOANY);
     delete m;
   }
   {
@@ -4678,9 +4676,11 @@ void testParseCHG() {
   }
 
   TEST_ASSERT(m);
-  std::string out = MolToMolBlock(*m);
-  // There are now dative bonds in the molecule, so the mol block will
-  // be forced to V3000.
+  // Write it out in V3000 format, which makes counting the different charges
+  // easier but is really because of a change that caused it always to write
+  // V3000 but which is no longer extant.
+  bool forceV3000(true);
+  std::string out = MolToMolBlock(*m, true, -1, true, forceV3000);
   std::regex chg_all("CHG="), chg_m1("CHG=-1"), chg_p1("CHG=1"),
       chg_p4("CHG=4");
   TEST_ASSERT(