+++ Code/GraphMol/MolStandardize/MolStandardize.cpp
@@ -285,5 +285,17 @@ std::vector<std::string> enumerateTautomerSmiles(
   return res.smiles();
 }
 
+void disconnectOrganometallics(
+    RWMol &mol, RDKit::MolStandardize::MetalDisconnectorOptions mdo) {
+  RDKit::MolStandardize::MetalDisconnector md(mdo);
+  md.disconnect(mol);
+}
+
+ROMol *disconnectOrganometallics(
+    const ROMol &mol, RDKit::MolStandardize::MetalDisconnectorOptions mdo) {
+  RDKit::MolStandardize::MetalDisconnector md(mdo);
+  return md.disconnect(mol);
+}
+
 }  // end of namespace MolStandardize
 }  // namespace RDKit